fix: use a lambda for repetitive shit
In read_from of TableMeta

I read from file to a certain delimiter, write all those to the raw
buffer of a string stream, then write that to a variable, then clear the
string, move the file pointer up by 1 position; rinse and repeat for all
variables needed.

That's a lot of repeating code.

commit-by: Huy Nguyen