free page impl, fix file write bug
I had a small undefined vtable bug, that is already fixed.

There's a bug where after the 3rd unit test, with the old API of
taking in the path, the file gets wiped out, and only the 4096th byte
and those of the first free page format gets written in. Changing the
API to take in the stream(s) itself fixed it.

commit-by: Huy Nguyen