import std in test, default value for use-import-std
The test files in dbfile/internal/test now use import std. Currently I'm
forcing a FetchContent (or in this case, CPMAddPackage) for GoogleTest
because there's a good chance that if you use a preinstalled GoogleTest
binary, it will have conflicts with all the modules, unless the
GoogleTest binary is built with the same flags you use for this project
(primarily, -stdlib=libc++ on Linux).
Later on, I may add the ability to use the default GoogleTest install.
But, that will probably have to wait until libstdc++ is updated to
version 15.

commit-by: Huy Nguyen