refactor(token): unit-test the tokenize interface
Instead of all the small things (from building a TokenizerData and a
Tokenizer functor, to calling it repeatedly). Just either the returned
vector, or the error code.

With this, the only error code to never be returned is
ParseError::ErrType::Done. I should have named things better. Done isn't
really an error.

Commit-by: Huy Nguyen